requirements.txt
starlette
uvicorn
httpx
pydantic
mcp
python-dotenv
